requirements.txt
browser-use>=0.7.10
playwright>=1.55.0
python-dotenv>=1.1.1
pandas>=2.0.0
openpyxl>=3.1.0
google-api-python-client>=2.184.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
twilio>=8.0.0
requests>=2.32.5
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0
loguru>=0.7.0
aiohttp>=3.8.0